requirements.txt
Flask
numpy
Pillow
tensorflow
pymysql
python-dotenv
gunicorn
pyjwt